loader loads on initial load
