+++ web/package.json
@@ -36,6 +36,9 @@
       "prettier/prettier": "warn"
     }
   },
+  "eslintIgnore":[
+    "build"
+  ],
   "browserslist": {
     "production": [
       ">0.2%",
